Category

See merge request wniemir1/bd2-z07!4

+++ src/main/java/warehouse/warehousemanagementsystem/category/CategoryService.java
@@ -1,8 +1,13 @@
 package warehouse.warehousemanagementsystem.category;
 
 import org.springframework.stereotype.Service;
+import warehouse.warehousemanagementsystem.exception.BadRequestException;
+import warehouse.warehousemanagementsystem.exception.ConflictException;
+import warehouse.warehousemanagementsystem.exception.DatabaseException;
+import warehouse.warehousemanagementsystem.exception.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoryService {
@@ -18,16 +23,33 @@ public class CategoryService {
         return categoryDao.getAllCategories();
     }
 
-    public Category getCategoryById(Long id) {
-        return categoryDao.getCategoryById(id);
+    public void addCategory(Category category) {
+        if (categoryDao.getCategoryByName(category.name()).isPresent()) {
+            throw new ConflictException("Category already exists");
+        }
+        if (category.name().isEmpty()) {
+            throw new BadRequestException("Category name cannot be empty");
+        }
+        if (categoryDao.addCategory(category) != 1) {
+            throw new DatabaseException("Failed to add category");
+        }
     }
 
-    public int addCategory(Category category) {
-        return categoryDao.addCategory(category);
-    }
+    public void deleteCategory(Long id) {
+        Optional<Category> category = categoryDao.getCategoryById(id);
+        int result;
+        if (category.isEmpty()) {
+            throw new NotFoundException("Category not found");
+        }
+        try {
+            result = categoryDao.deleteCategory(id);
+        } catch (Exception e) {
+            throw new ConflictException("Category is used as foreign key");
+        }
 
-    public int deleteCategory(Long id) {
-        return categoryDao.deleteCategory(id);
+        if (result != 1) {
+            throw new DatabaseException("Failed to delete category");
+        }
     }
